+++ BFKitDemo/BFKitDemo/DetailViewController.h
@@ -11,6 +11,7 @@
 typedef NS_ENUM(NSInteger, DetailType) {
     DetailTypeBFApp = 0,
     DetailTypeBFButton,
+    DetailTypeBFCryptor,
     DetailTypeBFDataStructures,
     DetailTypeBFLog,
     DetailTypeBFPassword,